+++ src/FluxFlow.Components.Timers/Nodes/TimerThrottleNodeFactory.cs
@@ -1,4 +1,5 @@
 using FluxFlow.Components.Timers.Options;
+using FluxFlow.Components.Timers.Timing;
 using FluxFlow.Engine.Runtime;
 using System.Reflection;
 
@@ -20,9 +21,10 @@ internal static class TimerThrottleNodeFactory
 
     private static RuntimeNode CreateThrottleTyped<TInput>(
         RuntimeNodeFactoryContext context,
-        TimerThrottleSettings settings)
+        TimerThrottleSettings settings,
+        ITimerClock clock)
     {
-        var node = new TimerThrottleNode<TInput>(settings);
+        var node = new TimerThrottleNode<TInput>(settings, clock);
 
         return context.CreateNode(node)
             .Input(TimerComponentPorts.Input, node.Input)